demonstration/wo AI renamed AccountSetup and its directory to match the Android naming convention Modified login page to use less absolute dimensions Moved navigation view to separate activity than login page
