+++ app/src/main/java/com/example/desparadosaeye/ui/MainActivity.kt
@@ -1,8 +1,7 @@
-package com.example.desparadosaeye
+package com.example.desparadosaeye.ui
 
 import android.os.Bundle
 import android.view.Menu
-import android.view.View
 import com.google.android.material.navigation.NavigationView
 import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
@@ -12,7 +11,7 @@ import androidx.navigation.ui.setupWithNavController
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
-import com.example.desparadosaeye.data.ApplicationModel
+import com.example.desparadosaeye.R
 
 class MainActivity : AppCompatActivity() {
 
@@ -30,7 +29,8 @@ class MainActivity : AppCompatActivity() {
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         appBarConfiguration = AppBarConfiguration(setOf(
-                R.id.nav_conversation, R.id.nav_account_management, R.id.nav_settings), drawerLayout)
+            R.id.nav_conversation, R.id.nav_account_management, R.id.nav_settings
+        ), drawerLayout)
         setupActionBarWithNavController(navController, appBarConfiguration)
         navView.setupWithNavController(navController)
     }
