+++ app/src/main/java/com/example/desparadosaeye/ui/login/LoginActivity.kt
@@ -1,0 +1,18 @@
+package com.example.desparadosaeye.ui.login
+
+import android.app.Activity
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+import com.example.desparadosaeye.R
+
+class LoginActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        setContentView(R.layout.activity_login)
+    }
+}
