+++ app/src/main/res/layout/activity_login.xml
@@ -1,0 +1,102 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="horizontal"
+    tools:context=".ui.login.LoginActivity">
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        android:layout_gravity="center"
+        tools:ignore="UselessParent">
+
+        <ImageView
+            android:id="@+id/imageView"
+            android:layout_width="141dp"
+            android:layout_height="124dp"
+            android:layout_gravity="center"
+            android:fadingEdge="horizontal|vertical"
+            android:foreground="@drawable/rounded_fg"
+            app:srcCompat="@drawable/cowboy_logo"
+            android:contentDescription="@string/nav_header_title" />
+        <!-- Email Prompt Field-->
+
+        <EditText
+            android:id="@+id/TextEmailAddress"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:ems="10"
+            android:hint="@string/prompt_email"
+            android:inputType="textEmailAddress"
+            tools:ignore="Autofill"
+            android:layout_marginTop="@dimen/login_fields_margin"/>
+
+        <EditText
+            android:id="@+id/TextPassword"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:ems="10"
+            android:hint="@string/prompt_password"
+            android:inputType="textPassword"
+            tools:ignore="Autofill"
+            android:layout_marginBottom="@dimen/login_fields_margin"/>
+
+        <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:orientation="horizontal">
+            <Button
+                android:id="@+id/buttonSignIn"
+                android:layout_width="@dimen/login_button"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center"
+                android:text="@string/log_in_button_width"
+                tools:ignore="ButtonStyle" />
+
+            <Button
+                android:id="@+id/buttonCreateAccount"
+                android:layout_width="@dimen/login_button"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center"
+                android:text="@string/sign_up_button"
+                tools:ignore="ButtonStyle" />
+        </LinearLayout>
+
+        <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:orientation="horizontal"
+            android:layout_margin="@dimen/auth_provider_button_margin">
+
+            <ImageButton
+                android:id="@+id/imageButtonGoogle"
+                android:layout_width="@dimen/auth_provider_button_size"
+                android:layout_height="@dimen/auth_provider_button_size"
+                android:layout_gravity="center"
+                android:foreground="@drawable/rounded_fg"
+                android:scaleType="fitXY"
+                app:srcCompat="@drawable/google_image"
+                android:layout_margin="@dimen/auth_provider_button_margin"
+                android:contentDescription="@string/sign_in_with_google_button" />
+
+            <ImageButton
+                android:id="@+id/imageButtonFacebook"
+                android:layout_width="@dimen/auth_provider_button_size"
+                android:layout_height="@dimen/auth_provider_button_size"
+                android:layout_gravity="center"
+                android:foreground="@drawable/rounded_fg"
+                android:scaleType="fitXY"
+                app:srcCompat="@drawable/facebook_image"
+                android:layout_margin="@dimen/auth_provider_button_margin"
+                android:contentDescription="@string/sign_in_with_facebook_button" />
+        </LinearLayout>
+    </LinearLayout>
+</LinearLayout>
