+++ app/src/main/res/layout/fragment_account_setup.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.accountSetup.AccountSetup">
+    tools:context=".ui.account_setup.AccountSetupFragment">
 
     <EditText
         android:id="@+id/editTextTextPersonName"
