+++ app/src/main/res/layout/fragment_settings.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.accountSetup.AccountSetup">
+    tools:context=".ui.account_setup.AccountSetupFragment">
 
     <TextView
         android:id="@+id/text_slideshow"
