+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -9,17 +9,7 @@
         android:id="@+id/nav_conversation"
         android:name="com.example.desparadosaeye.ui.conversation.ConversationFragment"
         android:label="@string/menu_conversation"
-        tools:layout="@layout/fragment_conversation" >
-        <action
-            android:id="@+id/action_nav_conversation_to_accountSetup"
-            app:destination="@id/nav_account_setup" />
-        <action
-            android:id="@+id/action_nav_conversation_to_nav_account_management"
-            app:destination="@id/nav_account_management" />
-        <action
-            android:id="@+id/action_nav_conversation_to_nav_settings"
-            app:destination="@id/nav_settings" />
-    </fragment>
+        tools:layout="@layout/fragment_conversation" />
 
     <fragment
         android:id="@+id/nav_account_management"
@@ -35,7 +25,7 @@
 
     <fragment
         android:id="@+id/nav_account_setup"
-        android:name="com.example.desparadosaeye.ui.accountSetup.AccountSetup"
+        android:name="com.example.desparadosaeye.ui.account_setup.AccountSetupFragment"
         android:label="@string/menu_account_setup"
-        tools:layout="@layout/account_setup_fragment" />
+        tools:layout="@layout/fragment_account_setup" />
 </navigation>
